+++ doc/CHANGES.rst
@@ -162,6 +162,9 @@ Features Added
 Bugs Fixed
 ++++++++++
 
+- LP #142451: If non-recursive ownership changes are made using 
+  ``changeOwnership``, do not touch any children.
+
 - LP #142563:  Fix ``AccessControl.User.NullUnrestrictedUserTests.__str__``.
 
 - LP #267820:  Fix bad except clause in the ``sequence_sort`` method of